+++ packages/components/src/ProgressCycle/ProgressCycle.stories.tsx
@@ -25,5 +25,9 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Basic: Story = {
-  render: args => <ProgressCycle {...args} />,
+  render: args => (
+    <div className="flex size-7 items-center justify-center bg-gray-500 p-4">
+      <ProgressCycle {...args} />
+    </div>
+  ),
 };