+++ themes/theme-b2b/src/components/Pagination.styles.ts
@@ -1,6 +1,6 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
-export const Pagination: ThemeComponent<'Pagination'> = cva([
+const BUTTON_STYLES = cva([
   'rounded-xs border-none leading-[48px] outline-hidden',
   'focus-visible:outline-outline-focus focus-visible:outline focus-visible:outline-offset-1',
   'disabled:text-text-base-disabled disabled:bg-bg-inverted-disabled disabled:cursor-not-allowed disabled:border disabled:border-solid',
@@ -9,3 +9,9 @@ export const Pagination: ThemeComponent<'Pagination'> = cva([
   'data-[selected=true]:border-0 data-[selected=true]:border-b-2 data-[selected=true]:border-solid data-[selected=true]:border-b-border-selected data-[selected=true]:bg-none data-[selected=true]:text-text-base  data-[selected=true]:font-bold',
   'text-text-base hover:bg-gray-100',
 ]);
+
+export const Pagination: ThemeComponent<'Pagination'> = {
+  navigationButton: BUTTON_STYLES,
+  pageButton: BUTTON_STYLES,
+  icon: cva('h-5 w-5'),
+};