+++ themes/theme-rui/src/components/Underlay.styles.ts
@@ -1,0 +1,9 @@
+import { ThemeComponent, cva } from '@marigold/system';
+
+export const Underlay: ThemeComponent<'Underlay'> = cva('', {
+  variants: {
+    variant: {
+      modal: ' bg-black/25 backdrop-blur-sm',
+    },
+  },
+});