+++ themes/theme-b2b/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @marigold/theme-b2b
 
+## 29.2.2
+
+### Patch Changes
+
+- Updated dependencies [40db199]
+- Updated dependencies [619b4b2]
+  - @marigold/components@11.2.1
+  - @marigold/system@11.2.1
+
 ## 29.2.1
 
 ### Patch Changes